Reorganized file structure for better readability
-Added different sizes of cloud PNG
-Moved media+stylesheets out of components
-deleted components/javascripts
-components only contains react components now
-consolidated all custom css into custom.css
-added tab favicon
