+++ Front_End/package.json
@@ -10,12 +10,13 @@
     "@testing-library/user-event": "^7.2.1",
     "bootstrap": "^4.5.2",
     "colorette": "^1.2.0",
+    "dependencies": "0.0.1",
     "google-map-react": "^2.0.8",
     "react": "^16.13.1",
     "react-bootstrap": "^1.3.0",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.2"
+    "react-scripts": "^3.4.3"
   },
   "scripts": {
     "start": "set PORT=3001 && react-scripts start",
