+++ Front_End/src/index.js
@@ -1,14 +1,14 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './components/stylesheets/index.css';
-import App from './App';
-import * as serviceWorker from './not_needed/serviceWorker';
+import React from "react";
+import ReactDOM from "react-dom";
+import "./stylesheets/custom.css";
+import App from "./App";
+import * as serviceWorker from "./not_needed/serviceWorker";
 
 ReactDOM.render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById("root")
 );
 
 // If you want your app to work offline and load faster, you can change
